+++ kradio3/plugins/oss-sound/oss-sound.h
@@ -135,7 +135,7 @@ RECEIVERS:
 
     bool noticeSoundStreamData(SoundStreamID id,
                                const SoundFormat &,
-                               const char *data, unsigned size,
+                               const char *data, size_t size,
                                const SoundMetaData &md
                               );
 
@@ -201,7 +201,7 @@ protected:
     SoundStreamID   m_PlaybackStreamID,
                     m_CaptureStreamID;
 
-    unsigned        m_BufferSize;
+    size_t          m_BufferSize;
     RingBuffer      m_PlaybackBuffer,
                     m_CaptureBuffer;
 
@@ -210,7 +210,7 @@ protected:
     time_t          m_CaptureStartTime;
 
 
-    unsigned        m_PlaybackSkipCount,
+    size_t          m_PlaybackSkipCount,
                     m_CaptureSkipCount;
 
     bool            m_EnablePlayback,
