+++ kradio3/plugins/recording/recording-monitor.cpp
@@ -283,7 +283,7 @@ bool RecordingMonitor::stopRecording(SoundStreamID id)
 }
 
 bool RecordingMonitor::noticeSoundStreamData(SoundStreamID id,
-    const SoundFormat &sf, const char *data, unsigned size,
+    const SoundFormat &sf, const char *data, size_t size,
     const SoundMetaData &md
 )
 {
