+++ kradio3/plugins/recording/recording.h
@@ -110,7 +110,7 @@ protected:
 
     bool startRecording(SoundStreamID id);
     bool startRecordingWithFormat(SoundStreamID id, const SoundFormat &sf, SoundFormat &real_format);
-    bool noticeSoundStreamData(SoundStreamID id, const SoundFormat &sf, const char *data, unsigned size, const SoundMetaData &md);
+    bool noticeSoundStreamData(SoundStreamID id, const SoundFormat &sf, const char *data, size_t size, const SoundMetaData &md);
     bool stopRecording(SoundStreamID id);
     bool isRecordingRunning(SoundStreamID id, bool &b) const;
